Data models (#39)
* Add models for event types

* Remove room_code from DisconnectData

* Add username to DisconnectData

* Add error event type

* Rename ConnectionData to ConnectData

* Add validator to Event

* Add status_code to Event

* Remove old event data classes

* Rename cls to self

Configures flake8 to ignore N805.

* Remove useless print statement

* Convert received data to Event

* Move ConnectionManager methods

* Add ConnectionManager.connect method

This method handles the connection type directly in the ConnectionManager.

* Convert ConnectionManager.connect to a static method

* Make _update_code_cache public

This also sets the line length for isort to 119.

* Make EventType a subclass of str

This helps with JSON serialization.

* Convert data type to Event

This also changes the docstrings of the modified functions.

* Update code cache tests

Updates the tests because _update_code_cache was made public.

* Match field order with docstring

* Add data attribute to RoomNotFoundError

* Edit Event validator to accept EventData

* Add difficulty validator

* Separate EventRequest and EventResponse

EventRequest is the data sent from the client to the server, EventResponse is the data sent from the server to the client. It allows to easily handle the status codes.

* Fix difficulty type annotation

* Fix EventRequest and EventResponse docstrings

* Remove useless docstrings

* Fix RuntimeError

This avoids that the client sends a message to itself after he disconnects.